runtimebp: GOMAXPROCS return two values instead of one
I thought runtime.GOMAXPROCS returns the new value set, but apparently I
was wrong and it actually returns the previous value. So make
runtimebp.GOMAXPROCS to return both old value (from the return value of
runtime.GOMAXPROCS) and new value to make it more useful for debugging.

This is a breaking change, but the impact should be minimal.