metricsbp: Split histogram into histogram and timing
We used to use go-kit's statsd implementation, which only provided
timing as histogram. We have since switched to influxstatsd
implementation, which actually provided both histogram and timing.

I looked into the code, looks like the only difference between them is
that for histograms it's writing to the wire with "|h" while for timing
it's writing to the wire with "|ms".