metricsbp: Split DefaultSampleRate into counter and histogram
In influxstatsd implementation, all histograms are not aggregated, and
forwarded to statsd as-is, while counters are aggregated (gauges are
always snapshotted without the conception of sample rate). This means we
usually need a different (lower) sample rate for histograms (including
timings) to avoid sending too much data to statsd.

Split DefaultSampleRate into DefaultCounterSampleRate and
DefaultHistogramSampleRate to resolve this problem.

Also for user convenience, treat zero value of the sample rate configs
as 1, and <0 as 0.

This is a breaking change.