metricsbp: Actually sample counters and histograms
go-kit's influxstatsd implementation doesn't actually do sampling. It
expects the user to do sampling manually, and report the sample rate to
them. Implement SampledCounter and SampledHistogram to actually do
sampling.

Also rename StatsdConfig.Default*SampleRate to *SampleRate, as the
semantic changed because of this change.

This is a breaking change.