thriftpool: Fix a bug on active connection counting
Previously we only decrease number of active connections when we
successfully returned the client to the pool, which means when there are
errors (usually when Close reported error, or when we try to open
a new client in Release and the opener returned an error), we don't
decrease the number, even though semantically the client is returned to
the pool. That leads to the pool reporting exhausted when even we tried
to return every client we got from the pool.