tracing: Implement opentracing interface
Make our Tracer to implement opentracing.Tracer, and make our Span to
implement opentracing.Span and opentracing.SpanContext, so that our
tracing implementation can be considered as an opentracing
implementation and users can start migrating to opentracing APIs.

Currently some of the APIs are implemented with no-ops, but all APIs we
care about are implemented.

The only breaking change is Span.SetTag, which now returns self for
chaining (this is part of opentracing.Span API).