Add bazel to CI
Note that because of the lack of persistent volume on drone CI, we can't
reuse the cache, which is the biggest advantage of Bazel, so the Bazel
runs in CI would be much slower.

Also set test sizes and mark three tests as flaky, and move external
dependencies into external.bzl.