Revert "metricsbp: Add total counter to span hook"
This reverts commit ee15a9795136a9ade6ccb025ee9dbfcf9d4054fc.

The lack of atomic metrics writing between success/fail and total
counters caused the current implementation not really useful in real
world scenarios. When using success/total as the success rate in
dashboards, we can often see it to be above 1. Reverting the change to
discourage that usage and still use success/(success+fail) instead. We
might want to revisit this idea in the future when we can make sure that
success/fail + total can be counted atomically.