httpbp: refactor to remove go-kit v0
The intention here is to replace go-kit as the HTTP/middleware
framework for HTTP services with a micro-framework that makes
it easier to work with the framework already provided by
"net/http" and add Middleware support.

This is the initial attempt and we will refine this further in
later commits.