thriftclient: add middleware for thrift.TClient (#134)
This is a breaking change that changes the API for how you
set up a Thrift client pool.  Specifically, we have removed
`MonitoredClient` entirely and replaced `NewMonitoredClientPool`
with `NewBaseplateClientPool` as the "golden path" option for
creating a Thrift connection pool in a Baseplate service.

The intention behind using Middleware rather than the previous
approach is to trade off some having a potentially slightly higher
learning curve for having a system that is easier to extend within
individual services to make it easy to test new ideas and customize
your usage.