Merge thriftclient into thriftbp package
Sometimes the line between thrift client and server is blurry. Also in
both upstream apache thrift package, and stdlib http package, they put
client and server code in the same package. So we follow this pattern.

Most changes in this commit are moving files around and renames.

This is a breaking change.