metricsbp: Add total counter to span hooks
We only need 2 of fail, success, and total. Before this change we use
fail and success, and eventually I think we would want to use fail and
total instead, but keep success for now for backward compatibility.

The reason success and total won't work is that there's no way to
guarantee that the success and total counter gets reported atomically,
so it would lead to situations that success reported a slightly higher
rate than total, which screws up the rate calculating. Although
theoretically this is also an issue with fail+total (or fail+success),
they are much more tolerable than success+total.