thriftbp: ClientPool implement TClient
Update `thriftbp.ClientPool` to implement `TClient`.  It pulls
a client from it's pool to make the actual `Call` and releases
it after the call finishes.  This allows you to pass a
`ClientPool` in to a thrift client as the `TClient` and avoid
the boilerplate of getting a client from a pool and releasing
it before each call.

This also allows us to trim down the exported interface in
`thriftbp`:

1. Don't export `ttlClient`.
2. Update `ttlClient` to allow you to keep a client in a pool
   indefinetly and configure it using the `ClientPoolConfig`.
3. Remove all the Factory methods that were used to create a
   ClientPool.
4. Don't export `getClient` or `releaseClient`.