tracing: initialize new top-level span if no headers are set
Check if any headers are set on the tracing.Headers object and
if *none* are, initialize a new, top-level server span instead.

Adds a new StartTopLevelServerSpan function to do this.

This patch also pulls the logic around parsing the different
values set in tracing.Headers into methods on the Headers object
itself (and adds tests for them) rather than hard-coding them in
StartSpanFromHeaders.