metricsbp: Add CounterWithRate/HistogramWithRate/TimingWithRate
Technically we can override the globally set sample rate by using
metricsbp.M.Statsd.NewCounter(name, rate) directly, but the problem with
that is that it only works if rate >= 1, because the implementation in
upstream doesn't actually do the sampling.

So add new functions to allow easier overriding of the globally set
sample rates.