metricsbp: Unexport Statsd.Statsd
The reason we exported that before is for users to override the globally
set sample rate, but since now we provided direct API to do that,
there's no need to export it.

Unexporting it also gives us the flexibility to switch the underlying
implementation without causing breaking changes in the future.

This is a breaking change, though.