thrifttest: add 'NewBaseplateServer'
This is inspired by httptest.NewServer and allows you to create
a new Thrift server listening using the local loopback interface
as well as a ClientPool for creating a Thrift client to talk to
the server.

This can be used to test your thrift service in automated testing.

It _can_ also be used to create fake services for testing your
integration, although you might also be served by using the Mocks
available in 'thrifttest' for that.