secrets: add 'NewTestSecrets'
This allows you to create a new secrets.Store using a MockFilewatcher
with the secrets you pass into it rather than having to create a
temporary directory and writing the secrets JSON yourself in tests.

This change also exports some additional parts of the secrets
package:

 * Encoding is now exported.  This is a part of the GenericSecret
   struct which is used by NewTestSecrets but was not previously
   exported.
 * Export the strings that define the three secret types.

Both of these are a part of the Baseplate spec so exporting them
should be reasonable and they are unlikely to change.