thriftbp: wrap the ClientPool Call method rather than wrapping the clients it creates
This is mainly so we apply middleware before getting a client from the pool, this lets us do things like use a new client if we trigger a retry due to a bad client in middleware.  It also has the side effect of only calling Wrap once rather than every time a new Client object is built.