httpbp: Add [en|de]codeEdgeContextHeader
This streamlines the encoding/decoding of edge context header so that
devs no longer need to think "is it standard base64 or url safe
base64?", also avoids the need of creating the header struct in
middleware while it's not really needed.

It's unexported as it's really just an internal thing. As a result, in
fixtures_test we still do base64 encoding directly as that's an external
test.