metricsbp: Switch back to use tags for sys stats
When we do aggregation for runtime.*.*.something gauges, we actually
convert the hostname and pid part into tags before doing the real
aggregation. We were doing that because we don't have tags in
Baseplate.py.

Since we already have tags in Baseplate.go, we really should just skip
that step.