metricsbp: Add Statsd.RuntimeGauge
This is the gauges we used in RunSysStats. Make it a public API so other
code can take advantage of it, too.

Also make thriftbp.ClientPool to report RuntimeGauges instead of plain
gauges. To make it non-breaking, currently we report both. The old,
plain gauges will be removed in a future version.