experiments: Speed up tests
Currently running tests for experiments with race detector on will take
close to 30s, which occasionally will cause the bazel tests to timeout
(as the timeout for small tests is 30s).

Change the number of users in TestCalculateBucketWithSeed and
TestCalculateBucket to a quarter what they are now, which reduces the
test running time to <10s.

As we have fewer users per bucket, the derivation of the number of users
per bucket is higher than before, so change the error margin from 0.1 to
0.15.

Also a few small improvements in those two tests to make them slightly
faster.