log: Add ctx to Wrapper API
This allows zap wrapper implementations to take advantage and don't lose
trace id when using Wrapper API to do logging.

This is a breaking change, but since log.Wrapper API is only supposed to
be used internally, for most users the breaking part is only that it can
no longer directly cast into thrift.Logger.