metricsbp: Buffer udp messages together
When using telegraf as the statsd collector, telegraf has a limit of how
many UDP messages can be buffered, and the default go-kit influxstatsd
implementation sends one UDP message per metric on the report interval,
so when we have a lot of metrics, telegraf could be dropping those UDP
messages, thus dropping metrics.

Add a configurable buffer size can help reduce the number of UDP
messages.

Eventually we probably want to upstream this to go-kit.