kafkabp: Add GroupID to ConsumerConfig
This adds the support for group consumers, for exactly-once message
queue consumer use cases.

Also some tweaks on the existing shared consumer interfaces. Most of the
tweaks are on the wording of the doc comment to add reference to the
newly added GroupID, but one notable change is that the error
return from ConsumeMessageFunc is removed. The reason is with added
group consumers, the error return could make users to think "If I
return an error in this function maybe the message won't be committed",
which is not true and also impossible to do, as in Apache Kafka message
committing is a watermark, not individually per message, so there's no
way to skip committing a single failed message.

Also there's no tests for group consumer, as sarama doesn't provide a
good mock implementation for it.