thriftbp: Add closed connections counter to client pools
The bug fixed in 6bfad2e highlights the fact that this is a part of the
code that has very limited external visibility, and when a bug happens
there it's very hard to debug (it took me almost 2 days to find it).

So adding a counter to provide external visibility, and if a similar bug
happens in the future, it would be more obvious that this counter
dropped while the actual failure rate remains the same.