Add ecinterface package
And switch everything else to use it instead of edgecontext package.

This is the first step of splitting edgecontext implementation to its
own repo. After this change, nothing in Baseplate.go depends on
edgecontext package any more, so we could move it to its own repo and
remote it from Baseplate.go in the next step.