metricsbp: Stop calling doWrite in Statsd.Close
We used to do that because the upstream implementation (go-kit statsd)
does not provide flushing ability in their ticker goroutine. We
reimplemented the ticker goroutine with the buffered writer change, and
added the ability to do one more flush before exiting, but forgot to
remove the extra flushing in Statsd.Close. This actually causes race
condition when Statsd.Close is called, because two goroutines are gonna
do the flushing at the same time, using the same writer.

While I'm here, also fix the potential deadlock issue in HandleShutdown
callback.