httpbp&thriftbp: Tweak default list for retryable codes
There's no guarantee that there's no side-effects in 500, 502, 504, and
507 errors. So be more conservative and don't mark them as retryable by
default.

Also tweak httpbp.ClientError.Retryable logic, to return 0 instead of -1
for codes not in the default list. This way users can implement their
own retrybp.Filter to mark certain codes not in the default list
retryable. On thriftbp the list is already customizable and also we
defer to the next filter instead of mark the error non-retryable when
it's not in the list, so this actually brings us feature parity (kinda)
between thriftbp and httpbp.

The current codes in default retryable list for both httpbp and thriftbp
are:

* 425 (too early)
* 429 (too many requests)
* 503 (service unavailable)