Add internal/limitreader package
This implements something similar to io.LimitReader, but reports error
when it's read beyond the limit.

It's under internal as this is a special case that we don't want to make
it part of our public API. We can move it out of internal in the future
if there are wider demands for it.

Switch filewatcher to use that instead of io.LimitReader. It could also
be used in https://github.com/reddit/baseplate.go/pull/385.