thriftbp: Don't suppress baseplate.Error in range [500, 600)
To follow the Baseplate spec update, don't suppress baseplate.Error when
the code is in range [500, 600).

Also remove thriftbp.BaseplateErrorCode. We added an unexported
baseplateError which is a super set of BaseplateErrorCode, and
maintaining an exported version of that is not really in our best
interest. I did an internal code search and didn't see any service using
BaseplateErrorCode so I'll skip the deprecation process and just remove
it directly.