baseplate: Implement two-step New
Remove NewArgs.ConfigPath in favor of two-step New, and change
DecodeConfigYAML into ParseConfigYAML to support two-step New.

Two-step New gives service code a chance to override some of the parsed
config before using them to initialize Baseplate packages. For example
Config.Sentry.BeforeSend cannot be deserialized from yaml so this is the
only chance for service code to set it.