metricsbp: Wait for flushing to finish in Statsd.Close
That was removed during the fix of the double-flush bug.

I tried to write a test for it but can't get it to pass reliably, likely
due to the caching on the system level (data written to an udp
connection is not always readable on the other end of the connection
immediately). No test is better than flaky test.