log: Add kv support to ZapWrapper
This would help people search for logs from certain upstream
dependencies, for example they can search for "from:thrift" for logs
produced from thrift's TSimpleServer.

This is a breaking change. But the majority of the internal usage are
from yaml config parsing, and there are only 19 direct usages of
log.ZapWrapper from internal code search, and the majority of the 19 are
from my team's monorepo.