thriftbp: Add background refresh for ttlClients
For every ttlClient that has a ttl, add a background goroutine to
refresh the connection when ttl hits.

This adds one goroutine per connection in the pool, until the connection
fails to refresh.

Both IsOpen and Call are protected by a locked state now. When ttlClient
is used, we always call IsOpen first then Call, and in theory these 2
calls might not be using the same connection (e.g. IsOpen is called
right before the connection expires), but even in that scenario, Call
would use a freshly refreshed connection, so things should still work.