Add panic recovery middleware to httpbp (#450)
While recovering from panics isn't necessary to avoid the server 
crashing like it is in thrift, the monitoring from baseplate.go is 
blind to panics in httpbp. 

This change adds an unexported middleware that is always 
applied to http requests that will recover from any panics 
and bubble those up the call as an HTTPError.