filewatcher: Fix flaky tests
When overwriting the file with new content for test, the overwriting was
done incorrectly, via overwriting in place, and can sometimes cause the
filewatcher to be triggered twice (once when opening and truncating the
file, once when the file is closed and writing is done). Add a helper
function to do the writing on a separated file and then use os.Rename
for an atomic move instead.