filewatcher: Fix flaky tests 2
Follow up on 846e891, the create-write-close writing can also trigger
filewatcher twice and can make the test flaky when the machine running
it is slow (e.g. when the CI is slow). Change all writes in filewatcher
tests to use the helper to do the atomic write and make sure that they
only trigger filewatcher once.