Add gRPC baseplate server implementation
This adds the neccessary code to have a baseplate service backed by a
gRPC server.

ServerConfig can be used to configure specific keepalive settings that
are used for point-to-point healthchecks, otherwise defaults are used.
It is also used to pass in the concrete gRPC service implementation
since gRPC service implementations do not share a common interface.

The baseplate server Close method is implemented using GracefulStop,
this may need to change to grpc.Server#Stop for an immediate shutdown.