metricsbp|prometheusbp: Add BoolString
Currently in Baseplate spec, we want different string values for bool
values in statsd vs. prometheus metric tags/labels: In statsd we want
capitalized "True"/"False" while in prometheus we want "true"/"false".
This can be confusing and lead to inconsistency, especially when
services are transitioning from statsd to prometheus metrics.

Provide metricsbp.BoolString and prometheusbp.BoolString in hope to help
reduce the confusion. For devs it should be easier to:

* If I'm using metricsbp to emit a metric, use metricsbp.BoolString
* If I'm using prometheus to emit a metric, use prometheusbp.BoolString