thriftbp: Deprecate ClientPoolConfig.InitialConnectionsFallback
Currently ClientPoolConfig.InitialConnectionsFallback is default to
false, due to how zero values in go work. Deprecate it to make it
default to true, because cascading failures during an outage is in
general much worse.

Add RequiredInitialConnections to allow the current default behavior
(that could cause cascading failures) for services that absolutely need
it.

Also instead of fallback to 0 connections, return whatever connections
we already established with the pool. This requires a small behavior
change inside clientpool package, that clientpool.NewChannelPool now
could return both non-nil Pool and error. An internal code search shows
that there are 2 services using that directly (we probably should have
made that an internal package to avoid that, but that ship has already
sailed). I'll send out separated PRs to handle the both non-nil Pool and
error case for the 2 services to avoid connection leaks when they
upgrade Baseplate.go to this version.