filewatcher: Change to pull+push mode
Currently filewatcher only work on push mode, by using fsnotify to
monitor the parent directory. This might break when the whole parent
directory is remounted. Add pull mode, that checks the file mtime with a
ticket, and force reload the file if the mtime is after the last parsed
mtime.