thriftbp: Deprecate AbandonCanceledRequests
And also remove it from BaseplateDefaultProcessorMiddlewares.

There are two places in thrift code checks for ErrAbandonRequest. The
most important one is in compiler generated Process function, which
happens before all the middlewares, so doing the "translation" in a
middleware will have no effect there. The other, minor one is in
TSimpleServer to close the client connection directly if it's an
ErrAbandonRequest, but if we didn't return ErrAbandonRequest the Process
function will try to write the error returned to the client and that
will fail, so TSimpleServer will close the connection anyways, and
having this middleware will not really do anything extra differently.

Services should do the context.Canceled -> thrift.ErrAbandonRequest
translation in their endpoint handler code instead.