thriftbp: Fix IDL exceptions in client middlewares
In client middlewares, any error defined in thrift IDL is actually not
returned via the error return, but written as part of the result
TStruct. As a result, none of the current client middlewares actually
get the exceptions defined in IDL correctly (this includes
baseplate.Error).

Add a helper function to get those errors out of the result TStruct via
reflection.