upgrade log to panic (#542)
Currently, if the `admin.Serve()` crashes, the resulting error is logged with a warning and the process keeps chugging along without everything it provides (like metrics). Upgrading the log to a Panicw will cause the go routine to panic and crash the program meaning the software can never be in a state where the metrics server is not also running successfully.