runtimebp: Extend GOMAXPROCS logic to account for additional scenarios
On discussion with kevlar, we came to the conclusion that there should be
4 possible settings for GOMAXPROCS which should cover all needs and
ensure adequate automation.  These are:
1. GOMAXPROCS if provided,
2. limit if known (with configurable multiplier),
3. fallback to a known environment variable Infrared can provide,
4. exactly 2 (for minimum parallelism)

By providing this hierarchy, we give maximum flexibility for devs,
while also providing safe and sane guardrails to work with, and
never end up in a scenario where bp.go tries to use physical core count.