Move GlobalRegistry from internalv2compat to internal/prometheusbpint
This is not really needed outside of Baseplate.go (unlike other stuff in
internalv2compat) so we can move it to an actual internal package to
stop fighting with staticcheck.

Also rename internal/prometheusbp to internal/prometheusbpint to avoid
the need to rename import when using it.

Also some fixups from previous commits:

1. We missed moving redisprom's MaxSizeGauge to GlobalRegistry
2. We missed moving custom exporters to GlobalRegistry
3. We didn't actually register the PromActive with prometheus register
4. SpanHook.PromActive is not something people can initialize outside of
   Baseplate.go, so unexport it.
5. Also add enforcement of prometheusbpint.GlobalRegistry to linters.sh