kafkabp: Metrics cleanup
Remove statsd metrics, and some tweaks to prometheus metrics to better
align with prometheus naming convention (use plural form for units).

For rebalances, since success rebalances happen so rare (consumer
rebalances just happen so rare in general, not that we rarely success),
the vectored version will cause rate to not really report anything >0,
so split that into total and failures_total instead.

For the rack id failures, an internal code search shows that the http
one is not used by anything and the aws one is only used by 3 repos, so
I opted against keep both and deprecate the old one, and will notify the
repo owners instead.