thriftbp+httpbp: Rename the RecoverPanic middlewares
Because server middlewares are in most call stacks, having "panic" in
the function names will cause searching for actual panics in log
ingester much harder (you don't want to use "-RecoverPanic" either
because they might be part of the actual panic call stack as well).

For thriftbp, since RecoverPanic was exported, keep it as an alias to
newly renamed RecoverPanik, but don't use it by default.