Create a secrets.Wrap API for wrapping alternate secrets implementations
This change allows creating an API-compatible Store object that delegates the calls to
secret-getting functions to another store.

This is necessary because baseplate.go/v0 and edgecontext/v0 both hard-code *secrets.Store,
and so it is not only a breaking change to replace them with an interface, it would require
a lock-step change across two v0.x libraries, which is very likely to bite consumers.