clientpool: Follow up on 6f97656
The only reason the constructor in clientpool would return both non-nil
Pool and error is to avoid making breaking changes. Since we already
made a breaking change in 6f97656, it's better to always return nil Pool
when returning an error.

Also fix %w used in logging, and add doccomment to the newly added
functions in thriftbp.