kafkabp: Add ConsumePartitionFuncProvider to non-group consumers
Currently when using non-group consumers, they always consume all
partitions from the topic.

This newly added ConsumePartitionFuncProvider allows the services to
optionally only consume from the partitions they are interested in. This
allows hugely reduced network traffic between the consumers and the
brokers when the service only need a subset of the partitions (for
example, the service itself is already sharded and the sharding is
aligned on kafka partitions).

Co-Authored-By: Gopal Rajpurohit <gopal.rajpurohit@reddit.com>