thriftbp: Report request/response payload sizes from client side
We currently already have a middleware to report that from the server
side, but that's simulated by reconstruct THeaderProtocols and
read/write to them, which will not include the header size and will not
reflect compressions (zlib, etc.).

This client side reporting is done directly at the TTransport (TSocket)
level, so it will reflect the exact number actually read/written at that
level. If zlib is enabled in THeaderProtocol the size will reflect
compressed payload. The only thing it will not reflect is potential SSL
layer (e.g. when using TSSLSocket over TSocket, which we don't support
yet).