httpbp/thriftbp: Tracing middlewares v2 compat fix
For the following 3 tracing middlewares:

- thriftbp server
- thriftbp client
- httpbp client

The current behavior is to short-circuit to v2 compat middleware if
injected, or fallback to v0 tracing implementation if not injected.
Change the fallback logic to no-op, as the v0 tracing implementation is
not really working any more and it will spam logs.

For httpbp server tracing middleware, remove it from defaults, as the
current v2 compat implementation is to skip it when injected as the API
is not sufficient for the v2 compat implementation to work.

This replaces and closes
https://github.com/reddit/baseplate.go/pull/664.