Make headerbp client middleware safer (#684)
After some investigation, we figured out that the cause of the error we encountered was due to the service wiring up the default client middleware manually and using the `NewBaseplateThriftPool` constructor, resulting in the default middleware running twice. This means that on the first pass, the middleware would add the header to the request. The second pass would then see this and fail the request. I made two changes to make this safer:

1. Added an idempotency check to the client middleware so it only runs once per call. Add an error log and a metric when we detect this happening.
2. Changed the client middlewares to instead drop new baseplate headers and log an error message rather than failing the call.