[Faults] Require a strict method match when `m` is provided.
Do not allow matching when the configuration is for a specific method
and the real method is unknown.

The main idea here is that if someone wanted to inject a fault on the
`foobar` method, a request with an unknown method should not also have
that fault injected just because the primary address matched.

Note that this has already been merged into bpv2 and this PR is just
to get bpv0 on the same page.