[Faults] Add support for matching on the host/authority header. (#700)
* Add support for matching on the host/authority header.

The PRs adding this to bpv2 have already been approved, and this PR is
just to get bpv0 up-to-date with those.

This introduces support for a new `h` parameter in the fault
configuration header value. Specifically, if the `h` parameter is
present, then the value of it will be checked against the
host/authority header of the outgoing request before determining
whether the request matches for fault injection.

Note that this is _not_ supported for gRPC due to the authority header
not being exposed by client-middleware runtime. If you'd like more
info on the workaround for gRPC, I'm happy to DM information on what
we plan to do.

Fix lint error

* Use existing header instead of creating a new one.

* Update thriftbp/client_middlewares.go

Co-authored-by: Andrew Boyle <pacejackson@users.noreply.github.com>

* Update internal/faults/headers_test.go

Co-authored-by: Andrew Boyle <pacejackson@users.noreply.github.com>

---------

Co-authored-by: Andrew Boyle <pacejackson@users.noreply.github.com>